RAM-dietary calculation of string partitions
Before, recursion was utilized to obtain string sections shorter than
the limit, in a way that the desired first part was given as the first
element of a tuple, and the remainder of the string was the second
element. Since strings are immutable, this meant many copies of the
remainder of the string.

Now, all that is done is to take slices by the indices of newlines in
the string.