Attempting to use selenium-webdriver to render page
Up to this point, pages are being fetched using the node-fetch package.
While this provides a simple HTML page, it is often required that pages
be rendered before actually being displayed, as there are many SPAs on
the web today. At first, using selenium-webdriver was attempted to
render the page by waiting for a few seconds before ultimately fetching
the content. This generated a problem, however, from the combination with
the serverless framework. Puppeteer was tried next, but it requires a heavy
headless version of the Chrome browser. To keep this package small, as is
required by AWS Lambda, a way of fetching the page, after a specified render
time, is necessary, probably through a third-party service. The next commit
should contain the necessary changes for that.